+++ codegenDataprotector.ts
@@ -1,0 +1,26 @@
+import type { CodegenConfig } from '@graphql-codegen/cli'
+import * as dotenv from 'dotenv';
+
+dotenv.config();
+
+const config: CodegenConfig = {
+  schema: process.env.VITE_DATAPROTECTOR_SUBGRAPH_URL,
+  documents: ['src/**/*DpQuery.ts'],
+  ignoreNoDocuments: true,
+  generates: {
+    './src/graphql/dataprotector/': {
+      preset: 'client',
+      config: {
+        documentMode: 'string'
+      }
+    },
+    './src/graphql/dataprotector/schema.graphql': {
+      plugins: ['schema-ast'],
+      config: {
+        includeDirectives: true
+      }
+    }
+  }
+}
+ 
+export default config