+++ src/modules/addresses/address/workers/beneficiaryDeals/addressContributionTable.tsx
@@ -1,5 +1,5 @@
 import { PREVIEW_TABLE_LENGTH, TABLE_REFETCH_INTERVAL } from '@/config';
-import { execute } from '@/graphql/execute';
+import { execute } from '@/graphql/poco/execute';
 import { useQuery } from '@tanstack/react-query';
 import { LoaderCircle } from 'lucide-react';
 import { DataTable } from '@/components/DataTable';
@@ -87,7 +87,6 @@ export function AddressContributionTable({
     addressAddress,
     currentPage: currentPage - 1,
   });
-  console.log('AddressContributionTable', contribution);
 
   return (
     <div className="space-y-6">