+++ src/modules/datasets/dataset/schema/datasetSchemaDpQuery.ts
@@ -1,0 +1,12 @@
+import { graphql } from '@/graphql/dataprotector/gql';
+
+export const datasetSchemaQuery = graphql(`
+  query DatasetSchema($datasetAddress: ID!) {
+    protectedData(id: $datasetAddress) {
+      schema {
+        path
+        type
+      }
+    }
+  }
+`);