+++ src/modules/datasets/dataset/schema/datasetsSchemaDpQuery.ts
@@ -1,0 +1,13 @@
+import { graphql } from '@/graphql/dataprotector/gql';
+
+export const datasetsSchemaQuery = graphql(`
+  query datasetsSchema($datasetIds: [Bytes!]!) {
+    protectedDatas(where: { id_in: $datasetIds }) {
+      address: id
+      schema {
+        path
+        type
+      }
+    }
+  }
+`);